// src/enums/Categories.ts
export enum Categories {
  Arts_And_Litterature = 'arts_and_literature',
  Film_And_TV = 'film_and_tv',
  Food_And_Drink = 'food_and_drink',
  General_Knowledge = 'general_knowledge',
  Geography = 'geography',
  History = 'history',
  Music = 'music',
  Science = 'science',
  Society_And_Culture = 'society_and_culture',
  Sport_And_Leisure = 'sport_and_leisure',
}